reclassify a test
